feat: full gender and characteristics selection
The process includes selection of a gender, 5 characteristics, a loading analyzer (dummy functionality right now),  and a result with hard-coded names for boys and girls. The characteristics selection disabled checkboxes when more than 5 are intented to be selected. Clicking on the homepage link resets the selection process.

Routes check that the selection process has started. If it has not, then the user is redirected to the homepage.

* NO styling yet